backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-filter==24.3
psycopg[binary]==3.2.1
redis==5.2.1
django-redis==5.4.0
openpyxl==3.1.5
gunicorn==22.0.0
